LPD-20274 move getHighlightedDDMStructuresJSONArray to the JournalConfigurationDisplayContext file because is only used there.
