+++ modules/apps/journal/journal-web/src/main/java/com/liferay/journal/web/internal/display/context/JournalDisplayContext.java
@@ -104,7 +104,6 @@ import com.liferay.portal.kernel.search.filter.Filter;
 import com.liferay.portal.kernel.search.generic.BooleanQueryImpl;
 import com.liferay.portal.kernel.security.permission.ActionKeys;
 import com.liferay.portal.kernel.security.permission.PermissionChecker;
-import com.liferay.portal.kernel.service.GroupLocalServiceUtil;
 import com.liferay.portal.kernel.service.PortletPreferencesLocalServiceUtil;
 import com.liferay.portal.kernel.service.ResourcePermissionLocalService;
 import com.liferay.portal.kernel.service.RoleLocalService;
@@ -738,30 +737,6 @@ public class JournalDisplayContext {
 		}
 
 		return highlightedDDMStructuresList;
-	}
-
-	public JSONArray getHighlightedDDMStructuresJSONArray() throws Exception {
-		return JSONUtil.toJSONArray(
-			getHighlightedDDMStructures(),
-			ddmStructure -> JSONUtil.put(
-				"ddmStructureId", String.valueOf(ddmStructure.getStructureId())
-			).put(
-				"name", ddmStructure.getName(_themeDisplay.getLocale())
-			).put(
-				"scope",
-				() -> {
-					Group group = GroupLocalServiceUtil.fetchGroup(
-						ddmStructure.getGroupId());
-
-					if (group != null) {
-						return LanguageUtil.get(
-							_themeDisplay.getLocale(),
-							group.getScopeLabel(_themeDisplay));
-					}
-
-					return StringPool.BLANK;
-				}
-			));
 	}
 
 	public List<TabsItem> getInfoPanelTabsItems(boolean journalArticle) {
